Create AccountBalance
Write a program that calculates how much money is in the account after you make a certain number of installments. On each line you will receive the amount you need to deposit in the account until you receive the command "NoMoreMoney". For each amount received, the console must display "Increase:" + the amount and be added to the account. If you get a number less than 0, "Invalid operation!" Must be displayed on the console. and the program ends. When the program ends, "Total:" + the total amount in the account, formatted to the second decimal place, must be printed.